backend/pyproject.toml
[project]
name = "torale"
version = "0.1.0"
description = "Grounded search monitoring platform for AI-powered conditional automation"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "webwhen", email = "hello@webwhen.ai"}
]
maintainers = [
    {name = "webwhen", email = "hello@webwhen.ai"}
]
keywords = [
    "ai",
    "automation",
    "monitoring",
    "grounded-search",
    "llm",
    "notifications",
    "web-scraping",
    "task-scheduler",
    "claude",
    "background-tasks",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Framework :: FastAPI",
    "Typing :: Typed",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic[email]>=2.9.0",
    "pydantic-settings>=2.4.0",
"httpx>=0.27.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "clerk-backend-api>=1.0.0",
    "apscheduler>=3.11.0,<4.0",
    "openai>=1.45.0",
    "anthropic>=0.34.0",
    "composio>=0.8.0",
    "python-dotenv>=1.0.0",
    "novu-py>=1.3.0",
    "markdown>=3.10",
    "cron-descriptor>=1.4.0",
    "slowapi>=0.1.9",
    "pypika-tortoise>=0.1.6",
    "bcrypt>=4.0.0",
    "posthog>=7.8.3",
    "a2a-sdk>=0.3.22",
    "redis[hiredis]>=5.0.0",
]

[project.urls]
Homepage = "https://webwhen.ai"
Documentation = "https://github.com/prassanna-ravishankar/torale#readme"
Repository = "https://github.com/prassanna-ravishankar/torale"
"Bug Tracker" = "https://github.com/prassanna-ravishankar/torale/issues"
Changelog = "https://github.com/prassanna-ravishankar/torale/releases"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
]
research = [
    "langfuse>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/webwhen"]

[tool.ruff]
line-length = 100
target-version = "py311"
fix = true

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
norecursedirs = ["src", ".git", ".venv", "node_modules", "dist", "build"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "ruff>=0.14.4",
]
